webapp/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
google-generativeai==0.8.3
pandas==2.2.3
torch==2.3.1
scikit-learn==1.5.2
joblib==1.4.2
fasttext-wheel==0.9.2
python-multipart==0.0.9
jinja2==3.1.4
httpx==0.27.2
